pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "ai-digest"
version = "1.0.0"
description = "CLI tool to aggregate and summarize AI/ML news from RSS feeds"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "feedparser>=6.0.0",
    "pyyaml>=6.0",
    "openai>=1.0.0",
    "anthropic>=0.18.0",
]

[project.scripts]
ai-digest = "ai_digest.cli:main"

[tool.setuptools.packages.find]
where = ["."]
